// types.ts
export interface WorkdaySettings {
  workStart: string;
  workEnd: string;
  gracePeriod: number;
  workHoursPerDay: number;
  workingDays: string[];
}

export interface PublicHoliday {
  id: string;
  name: string;
  date: string;
  isRecurring: boolean;
}